Fit a broken stick model to irregular data

The broken stick model models an irregularly observed series of measurement by scaling them onto a user-specified set of 'ideal' ages. The model codes age by a series of linear B-splines. Differences between persons are expressed by a random effect pertaining to each knot. On the individual level, each modeled growth curve connect straight lines that join at the chosen break ages, and hence look like a 'broken stick'.

brokenstick(
  y,
  x,
  subjid,
  k = NULL,
  knots = NULL,
  boundary = NULL,
  control = lmerControl(check.nobs.vs.nRE = "warning"),
  na.action = na.exclude,
  ...
)

Arguments

y
a vector containing the measurements to be analyzed

x
a vector of length length(y) with the explanatory variable on which the break points should be defined. In longitudinal data, this is usually age.

subjid
a vector length length(y) containing the subject identification

k
optional, scalar indicating the number of internal knots. If specified, then k internal knots are placed at equidense quantiles of x. For example, specifying k = 1 puts a knot at the 50th quantile (median), specifying k = 3 puts knots at the 25th, 50th and 75th quantiles of x, and so on. If both k and knots are specified, then k take precendence. Note that knots specified via k are data-dependent and do not transfer well to other data sets. Use knots to specify knots that are independent of the data x.

knots
optional, numerical vector with the locations of the breaks to be placed on the values of x. Be careful with values outside the range of the data since this extends the boundary knots (see below) beyond the data range.

boundary
optional, numerical vector of length 2 with the minimum and maximum knot. This boundary setting is passed to splines::bs() as the Boundary.knots argument. If not specified, then the range of x is taken. If knots is specified, then the boundary range is extended to include at least range{knots}.

control
A function to control fitting of lmer(). The default is set to lmerControl(check.nobs.vs.nRE = "warning"), which turn fatal errors with respect the number of parameters into warnings.

na.action
The function to call for the na.action argument in lmer(). The default is na.exclude.

...
Additional arguments passed down to lmer() (e.g. to specify additional lmer() options.

Value

A fitted model of class brokenstick, which extends the class lmerMod

Details

Relations over time are modeled by the variance-covariance parameters of the random effects. Currently, this matrix is estimated as unstructured by lmer() from the lme4 package. This estimate may be unstable if the number of children is small relative to the number of specified knots.

This function can be time consuming for data sets with thousands of children.

Note

The storeX and degree arguments have been deprecated in version 0.54.

Examples

library(mice)

#> 
#> Attaching package: ‘mice’

#> The following objects are masked from ‘package:base’:
#> 
#>     cbind, rbind

data <- tbc[tbc$id < 1000 & tbc$age < 2.5, ]
fit <- brokenstick(
  y = data$hgt.z, x = data$age, subjid = data$id,
  knots = c(0, 1, 2)
)