Obtain the X model matrix from a broken stick model

Obtain the X model matrix from a broken stick model

get_X(object, ids = NULL)

Arguments

object
An object of class brokenstick

ids
A vector specifying the id's of the persons. If omitted, all id's are included.

Value

A matrix with number of columns equal to the number of knots. The result is NULL if object is not of class brokenstick.

Examples

get_X(fit_200, ids = c(10001, 10002))

#>            x1         x2         x3        x4     x5     x6     x7     x8
#> 1  1.00000000 0.00000000 0.00000000 0.0000000 0.0000 0.0000 0.0000 0.0000
#> 2  0.01440576 0.98559424 0.00000000 0.0000000 0.0000 0.0000 0.0000 0.0000
#> 3  0.00000000 0.09472422 0.90527578 0.0000000 0.0000 0.0000 0.0000 0.0000
#> 4  0.00000000 0.00000000 0.00000000 0.9816000 0.0184 0.0000 0.0000 0.0000
#> 5  0.00000000 0.00000000 0.00000000 0.0000000 0.9848 0.0152 0.0000 0.0000
#> 6  0.00000000 0.00000000 0.00000000 0.0000000 0.0000 0.9884 0.0116 0.0000
#> 7  0.00000000 0.00000000 0.00000000 0.0000000 0.0000 0.0000 0.9916 0.0084
#> 8  0.00000000 0.00000000 0.00000000 0.0000000 0.0000 0.0000 0.3020 0.6980
#> 9  0.00000000 0.00000000 0.00000000 0.0000000 0.0000 0.0000 0.0000 0.0000
#> 10 0.00000000 0.00000000 0.00000000 0.0000000 0.0000 0.0000 0.0000 0.0000
#> 11 1.00000000 0.00000000 0.00000000 0.0000000 0.0000 0.0000 0.0000 0.0000
#> 12 0.00000000 0.65227818 0.34772182 0.0000000 0.0000 0.0000 0.0000 0.0000
#> 13 0.00000000 0.19304556 0.80695444 0.0000000 0.0000 0.0000 0.0000 0.0000
#> 14 0.00000000 0.00000000 0.04321729 0.9567827 0.0000 0.0000 0.0000 0.0000
#> 15 0.00000000 0.00000000 0.00000000 0.0176000 0.9824 0.0000 0.0000 0.0000
#> 16 0.00000000 0.00000000 0.00000000 0.0000000 0.0212 0.9788 0.0000 0.0000
#> 17 0.00000000 0.00000000 0.00000000 0.0000000 0.0000 0.0248 0.9752 0.0000
#> 18 0.00000000 0.00000000 0.00000000 0.0000000 0.0000 0.0000 0.0000 0.0000
#> 19 0.00000000 0.00000000 0.00000000 0.0000000 0.0000 0.0000 0.0000 0.0000
#>        x9    x10    x11
#> 1  0.0000 0.0000 0.0000
#> 2  0.0000 0.0000 0.0000
#> 3  0.0000 0.0000 0.0000
#> 4  0.0000 0.0000 0.0000
#> 5  0.0000 0.0000 0.0000
#> 6  0.0000 0.0000 0.0000
#> 7  0.0000 0.0000 0.0000
#> 8  0.0000 0.0000 0.0000
#> 9  0.8678 0.1322 0.0000
#> 10 0.0000 0.9904 0.0096
#> 11 0.0000 0.0000 0.0000
#> 12 0.0000 0.0000 0.0000
#> 13 0.0000 0.0000 0.0000
#> 14 0.0000 0.0000 0.0000
#> 15 0.0000 0.0000 0.0000
#> 16 0.0000 0.0000 0.0000
#> 17 0.0000 0.0000 0.0000
#> 18 0.9610 0.0390 0.0000
#> 19 0.0794 0.9206 0.0000