Predict growth curve according to the broken stick model

Calculation predictions (conditional means of the random effects) from the broken stick model. This function takes an object of class brokenstick, and returns predictions in one of several formats. We may calculate predictions for new persons, i.e., for persons who are not part of the fitted model, by specifying the x and y arguments.

# S3 method for brokenstick
predict(object, y, x, ids = NULL, at = "x", output = "long", ...)

Arguments

object
An object of class brokenstick or of class brokenstick_export.

y
A vector with measurements using the same response scale as the fitted model.

x
A vector with decimal ages of length length(y). If both y and x are not specified, the function calculates the predicted values for all persons on which object was fitted. If y is not specified, but x is, then broken stick estimates are obtained at x for all persons in object. Note that the at = "x" argument is needed to make predictions at the x values.

ids
A vector with one or more subject identification codes.

at
If at = "x" (the default) the function returns predictions at the x values in the data. Specify at = "knots" to obtain predictions at the knots of the model, or at = "both" to obtain both.

output
A string specifying the desired type of output. If output = "long" (the default), the result is cast into a data frame of the long format that represents each row represented as an observation. If output = "broad", the result is formed into a broad matrix where each row represents a person. This format is useful as input to a secondary data analysis that analyzes the smoothed data as repeated measured. The format can be used if at = "knots", and returns NULL otherwise. If output = "vector", the result is returned as a vector. This is the fastest method.

...
Additional arguments passed down to predict.merMod(), predict.brokenstick_export() and predict_all().

Value

If output == "long", a data frame with five columns named:

subjid

Person identification

x

The x values of the model, usually age

y

Response values (in scale defined by the model)

yhat

Predicted values

knot

Logical, TRUE indicates that the prediction is done at the location of the knot, or at a location specified by an isolated x argument, i.e., without a corresponding y argument.

If output == "broad", a numeric matrix of predicted values with length(slot(object, "knots")) columns. The number of rows will equal the number of subjid's in the original data. Subject without any valid outcome will have NA's everywhere.

If output == "vector", a numeric vector of predicted values corresponding to the column yhat.

Note

The original data are only present in the object of class brokenstick. Hence, any calculations involving observations of the fitted model require an object of class brokenstick. Predictions for new units can be done both with class brokenstick and brokenstick_export.

See also

Examples

# Estimates at each measured age for all persons in fitted model
class(fit_50)

#> [1] "brokenstick"

p <- predict(fit_50)
head(p)

#>   subjid      x      y       yhat  knot
#> 1  10001 0.0000  0.575  0.5117682 FALSE
#> 2  10001 0.0821  0.888  0.8762757 FALSE
#> 3  10001 0.1588  0.797  0.8107744 FALSE
#> 4  10001 0.2546  0.661  0.6530335 FALSE
#> 5  10001 0.5038  0.290  0.2564071 FALSE
#> 6  10001 0.7529 -0.398 -0.2521086 FALSE

# Estimates at knots, stored into the broad matrix
p <- predict(fit_50, output = "broad", at = "knots")
round(head(p), 2)

#>           0 0.0833 0.1667  0.25   0.5  0.75     1  1.25   1.5     2      3
#> 10001  0.51   0.88   0.80  0.66  0.26 -0.26  0.20  0.08 -0.08  0.06  -1.65
#> 10002 -0.18  -0.29  -0.27 -0.29 -0.26 -0.26 -0.30 -0.31 -0.52 -0.66   2.03
#> 10003  1.13   2.09   2.07  1.99  2.18  1.85  1.33  0.92  0.82  1.23 -10.49
#> 10004  0.09   0.52   0.36  0.88  0.78  0.52  0.23 -0.16 -0.36 -0.10  -4.95
#> 10005  1.23   1.83   1.35  0.93  0.91  0.97  1.69  2.48  2.29  1.85   7.47
#> 10006 -0.61   0.17  -0.08  0.53 -0.03 -0.17 -0.48  0.17  0.34  0.45   5.22

# Get these estimates as a vector, useful for programming
p <- predict(fit_50, output = "vector", at = "knots")
round(head(p), 2)

#> [1]  0.51  0.88  0.80  0.66  0.26 -0.26

# Obtain estimates at weeks 1-4 for all children, include old points
p <- predict(fit_50, x = round((1:4) * 7 / 365.25, 4))
head(p)

#>   subjid      x      y       yhat  knot
#> 1  10001 0.0000  0.575  0.5117682 FALSE
#> 2  10001 0.0821  0.888  0.8762757 FALSE
#> 3  10001 0.1588  0.797  0.8107744 FALSE
#> 4  10001 0.2546  0.661  0.6530335 FALSE
#> 5  10001 0.5038  0.290  0.2564071 FALSE
#> 6  10001 0.7529 -0.398 -0.2521086 FALSE

# Same, but now organised as broad matrix of new points only
p <- predict(fit_50, x = round((1:4) * 7 / 365.25, 4), output = "broad")

#> Warning: data length [202] is not a sub-multiple or multiple of the number of rows [51]

head(p)

#>          0.0192     0.0383     0.0575      0.0767
#> [1,]  0.5970123  0.6818125  0.7670567  0.85230081
#> [2,] -0.2052942 -0.2308557 -0.2565509 -0.28224614
#> [3,]  1.3549098  1.5748444  1.7959305  2.01701654
#> [4,]  0.1885569  0.2858509  0.3836542  0.48145762
#> [5,]  1.2251736  1.3637778  1.5016601  1.64026434
#> [6,]  1.7788685 -0.4319678 -0.2527700 -0.07263408

head(predict(fit_50, at = "knots", output = "broad"), 3)

#>                0     0.0833     0.1667       0.25        0.5       0.75
#> 10001  0.5117682  0.8816035  0.8033631  0.6603196  0.2643374 -0.2573963
#> 10002 -0.1795990 -0.2910789 -0.2709276 -0.2918563 -0.2620037 -0.2648078
#> 10003  1.1338237  2.0930149  2.0693730  1.9904651  2.1840112  1.8468527
#>                1       1.25         1.5           2          3
#> 10001  0.1984385  0.0811640 -0.08167399  0.05944318  -1.652566
#> 10002 -0.3023886 -0.3088352 -0.51841890 -0.65750137   2.031763
#> 10003  1.3306968  0.9184405  0.82471139  1.22691579 -10.491509

predict(fit_50, id = c(10001, 10003), at = "knots", output = "broad")

#>               0    0.0833    0.1667      0.25       0.5       0.75         1
#> 10001 0.5117682 0.8816035 0.8033631 0.6603196 0.2643374 -0.2573963 0.1984385
#> 10003 1.1338237 2.0930149 2.0693730 1.9904651 2.1840112  1.8468527 1.3306968
#>            1.25         1.5          2          3
#> 10001 0.0811640 -0.08167399 0.05944318  -1.652566
#> 10003 0.9184405  0.82471139 1.22691579 -10.491509