Predict growth curve according to the broken stick model

Predict growth curve according to the broken stick model

# S3 method for brokenstick_export
predict(object, y, x, at = "x", output = "long", subjid = NA, ...)

Arguments

object
An object of class brokenstick or of class brokenstick_export.

y
A vector with measurements using the same response scale as the fitted model.

x
A vector with decimal ages of length length(y). If both y and x are not specified, the function calculates the predicted values for all persons on which object was fitted. If y is not specified, but x is, then broken stick estimates are obtained at x for all persons in object. Note that the at = "x" argument is needed to make predictions at the x values.

at
If at = "x" (the default) the function returns predictions at the x values in the data. Specify at = "knots" to obtain predictions at the knots of the model, or at = "both" to obtain both.

output
A string specifying the desired type of output. If output = "long" (the default), the result is cast into a data frame of the long format that represents each row represented as an observation. If output = "broad", the result is formed into a broad matrix where each row represents a person. This format is useful as input to a secondary data analysis that analyzes the smoothed data as repeated measured. The format can be used if at = "knots", and returns NULL otherwise. If output = "vector", the result is returned as a vector. This is the fastest method.

subjid
A subject identification

...
Additional arguments passed down to predict.merMod(), predict.brokenstick_export() and predict_all().

See also

Examples

exp <- export(fit_50)

# no data predicts mean trajectory
p <- predict(exp)
head(p)

#>   subjid      x  y         yhat knot
#> 1     NA 0.0000 NA -0.082753698 TRUE
#> 2     NA 0.0833 NA  0.398328018 TRUE
#> 3     NA 0.1667 NA  0.311552960 TRUE
#> 4     NA 0.2500 NA  0.350124345 TRUE
#> 5     NA 0.5000 NA  0.228374241 TRUE
#> 6     NA 0.7500 NA -0.001771215 TRUE

# predict mean trajectory at weeks 1-4
predict(exp, x = round((1:4) * 7 / 365.25, 4))

#>   subjid      x  y       yhat  knot
#> 1     NA 0.0192 NA 0.02813188 FALSE
#> 2     NA 0.0383 NA 0.13843994 FALSE
#> 3     NA 0.0575 NA 0.24932552 FALSE
#> 4     NA 0.0767 NA 0.36021110 FALSE

# add data at each week
predict(exp, x = round((1:4) * 7 / 365.25, 4), y = c(1, 0.8, 0.9, 0.7))

#>   subjid      x   y      yhat  knot
#> 1     NA 0.0192 1.0 0.9148645 FALSE
#> 2     NA 0.0383 0.8 0.8690146 FALSE
#> 3     NA 0.0575 0.9 0.8229246 FALSE
#> 4     NA 0.0767 0.7 0.7768346 FALSE

# no data at weeks 2 and 3
predict(exp, x = round((1:4) * 7 / 365.25, 4), y = c(1, NA, NA, 0.7))

#>   subjid      x   y      yhat  knot
#> 1     NA 0.0192 1.0 0.9301203 FALSE
#> 2     NA 0.0383  NA 0.8717094 FALSE
#> 3     NA 0.0575  NA 0.8129926 FALSE
#> 4     NA 0.0767 0.7 0.7542759 FALSE

# estimates at standard knots
predict(exp,
  x = round((1:4) * 7 / 365.25, 4), y = c(1, NA, NA, 0.7),
  at = "knots"
)

#>    subjid      x  y         yhat knot
#> 1      NA 0.0000 NA  0.988837030 TRUE
#> 2      NA 0.0833 NA  0.734092008 TRUE
#> 3      NA 0.1667 NA  0.712966698 TRUE
#> 4      NA 0.2500 NA  0.646615968 TRUE
#> 5      NA 0.5000 NA  0.283644520 TRUE
#> 6      NA 0.7500 NA  0.137621699 TRUE
#> 7      NA 1.0000 NA  0.146645329 TRUE
#> 8      NA 1.2500 NA  0.060475618 TRUE
#> 9      NA 1.5000 NA -0.002533962 TRUE
#> 10     NA 2.0000 NA -0.181272979 TRUE
#> 11     NA 3.0000 NA  1.689911149 TRUE

# leaving out the missing data produces the same result
predict(exp,
  x = round(c(1, 4) * 7 / 365.25, 4), y = c(1, 0.7),
  at = "knots"
)

#>    subjid      x  y         yhat knot
#> 1      NA 0.0000 NA  0.988837030 TRUE
#> 2      NA 0.0833 NA  0.734092008 TRUE
#> 3      NA 0.1667 NA  0.712966698 TRUE
#> 4      NA 0.2500 NA  0.646615968 TRUE
#> 5      NA 0.5000 NA  0.283644520 TRUE
#> 6      NA 0.7500 NA  0.137621699 TRUE
#> 7      NA 1.0000 NA  0.146645329 TRUE
#> 8      NA 1.2500 NA  0.060475618 TRUE
#> 9      NA 1.5000 NA -0.002533962 TRUE
#> 10     NA 2.0000 NA -0.181272979 TRUE
#> 11     NA 3.0000 NA  1.689911149 TRUE