Plot observed and fitted trajectories from fitted brokenstick model

This function plot the observed and fitted trajectories from brokenstick model. Actual plotting may be done by gglot.

plot_trajectory(
  x,
  data,
  color_y = c("blue", "grey"),
  size_y = 2,
  color_yhat = c("red", "grey"),
  size_yhat = 2,
  ncol = 3L,
  xlab = "Age (years)",
  ylab = "Length (SDS)",
  zband = FALSE,
  zband_range = NULL,
  xlim = NULL,
  ylim = NULL,
  scales = "fixed",
  theme = ggplot2::theme_light()
)

Arguments

x
An object of class brokenstick.

data
A data.frame with columns names x$names, ".source" and ".pred".

color_y
A character vector with two elements specifying the symbol and line color of the measured data points

size_y
Dot size of measured data points

color_yhat
A character vector with two elements specifying the symbol and line color of the predicted data points

size_yhat
Dot size of predicted data points

ncol
Number of columns in plot

xlab
The label of the x-axis

ylab
The label of the y-axis

zband
A logical indicating whether the Z-score band should be added to the plot. The default is FALSE.

zband_range
a vector specifying the range (min, max) that the superposed growth standard should span on the x-axis. The default is the entire data range.

xlim
Vector of length 2 with range of x-axis

ylim
Vector of length 2 with range of y-axis

scales
Axis scaling, e.g. "fixed", "free", and so on

theme
Plotting theme

Value

An object of class ggplot

See also

Examples

# plot first three cases
# plot(fit_200, brokenstick::smocc_200)