Predict from a brokenstick model

The predictions from a broken stick model coincide with the group-conditional means of the random effects. This function takes an object of class brokenstick and returns predictions in one of several formats. The user can calculate predictions for new persons, i.e., for persons who are not part of the fitted model, through the x and y arguments.

# S3 method for brokenstick
predict(
  object,
  new_data = NULL,
  type = "numeric",
  ...,
  x = NULL,
  y = NULL,
  group = NULL,
  strip_data = TRUE,
  shape = c("long", "wide", "vector")
)

Arguments

object
A brokenstick object.

new_data
A data frame or matrix of new predictors.

type
A single character. The type of predictions to generate. Valid options are:

"numeric" for numeric predictions.

...
Not used, but required for extensibility.

x
Optional. A numeric vector with values of the predictor. It could also be the special keyword x = "knots" replaces x by the positions of the knots.

y
Optional. A numeric vector with measurements.

group
A vector with group identifications

strip_data
A logical indicating whether the row with the observed data from new_data should be stripped from the return. The default is TRUE. Set to FALSE to infer which data points are extracted from new_data.

shape
A string: "long" (default), "wide" or "vector" specifying the shape of the return value. Note that use of "wide" with many unique values in x creates an unwieldy, large and sparse matrix.

Value

A tibble of predictions. If x, y and group are not specified, the number of rows in the tibble is guaranteed to be the same as the number of rows in new_data.

Details

By default, predict() calculates predictions for every row in new_data. It is possible to tailor the behavior through the x, y and group arguments. What exactly happens depends on which of these arguments is specified:

If the user specifies x, but no y and group, the function returns - for every group in new_data - predictions at x values. This method will use the data from new_data.

If the user specifies x and y but no group, the function forms a hypothetical new group with the x and y values. This method uses no information from new_data.

If the user specifies group, but no x or y, the function searches for the relevant data in new_data and limits its predictions to the specified groups. This is useful if prediction for only one or a few groups is needed.

If the user specifies x and group, but no y, the function will create new values for x in each group, search for the relevant data in new_data and limit prediction to locations x in those groups.

If the user specifies x, y and group, the functions assumes that these vectors form a data frame. The lengths of x, y and group must be the same. This procedure uses only information from new_data for groups with group values that match those on newdata.

As case 5, but now without a new_data argument. All data are specified through x, y and group. No matching to new_data attempted.

Examples

train <- smocc_200[1:1198, ]
test <- smocc_200[1199:1940, ]

# Fit
fit <- brokenstick(hgt.z ~ age | id, data = train, knots = 0:3)

# Predict, with preprocessing
predict(fit, test)

#>             .pred
#> 1    0.3189266651
#> 2    0.3254169175
#> 3    0.3321530793
#> 4    0.3422529307
#> 5    0.3624526336
#> 6    0.3874563528
#> 7    0.4056161192
#> 8    0.3855293149
#> 9    0.3556436436
#> 10   0.3221685477
#> 11  -0.2320679553
#> 12  -0.1786100919
#> 13  -0.1066822364
#> 14  -0.0490949032
#> 15   0.1666886633
#> 16   0.3537911852
#> 17   0.5237801145
#> 18   0.6455581325
#> 19   0.7780913346
#> 20   1.0042924141
#> 21   0.0745109091
#> 22   0.0547214963
#> 23   0.0413187726
#> 24   0.0145133254
#> 25  -0.0435496043
#> 26  -0.1150152576
#> 27  -0.1805670772
#> 28  -0.2087025416
#> 29  -0.2790503793
#> 30  -0.3554842651
#> 31   0.4884070339
#> 32   0.3841615137
#> 33   0.2321566305
#> 34   0.1104575226
#> 35  -0.3152513518
#> 36  -0.6757472753
#> 37  -1.0934869769
#> 38  -1.0475472747
#> 39  -0.9824443822
#> 40  -0.9063910954
#> 41  -0.2605141385
#> 42  -0.2851003993
#> 43  -0.3126700805
#> 44  -0.3471052235
#> 45  -0.4573480038
#> 46  -0.5606534327
#> 47  -0.6045984062
#> 48  -0.5592864578
#> 49  -0.4848774783
#> 50  -0.3575096550
#> 51  -1.5631191491
#> 52  -1.3724801375
#> 53  -1.2698420929
#> 54  -1.1331103674
#> 55  -0.6688294038
#> 56  -0.2093679658
#> 57   0.2216034802
#> 58   0.2120726647
#> 59   0.2003465489
#> 60   0.1819386696
#> 61  -0.2132146799
#> 62  -0.2907680721
#> 63  -0.4069464915
#> 64  -0.4650862576
#> 65  -0.7170589479
#> 66  -0.9883441518
#> 67  -1.2236462657
#> 68  -1.2126182954
#> 69  -1.1929918717
#> 70  -0.9578039689
#> 71   0.6786964739
#> 72   0.7139530319
#> 73   0.7613909287
#> 74   0.7993709568
#> 75   0.9322267784
#> 76   1.0461173449
#> 77   1.1440286507
#> 78   1.0835464015
#> 79   1.1286470779
#> 80  -0.8353706946
#> 81  -0.7857010662
#> 82  -0.7300042753
#> 83  -0.6604377490
#> 84  -0.4794776409
#> 85  -0.2846477974
#> 86  -0.1086466457
#> 87  -0.0900433280
#> 88  -0.0683792717
#> 89  -0.0010680596
#> 90   0.2143404671
#> 91   0.2065420080
#> 92   0.1960491376
#> 93   0.1855562671
#> 94   0.1582616604
#> 95   0.1330700090
#> 96   0.1048333651
#> 97   0.1053610047
#> 98   0.1059757016
#> 99   0.1487434602
#> 100 -0.6128135041
#> 101 -0.5042562377
#> 102 -0.4120486815
#> 103 -0.3198411252
#> 104  0.0487486628
#> 105  0.3252511131
#> 106  0.6281238290
#> 107  0.6658194491
#> 108  0.7557706391
#> 109  0.1875349267
#> 110  0.2494461249
#> 111  0.3267745837
#> 112  0.3886050633
#> 113  0.5897559186
#> 114  0.7908260552
#> 115  0.9919769105
#> 116  1.0010883769
#> 117  1.0085480237
#> 118  1.0312465566
#> 119 -1.6397302118
#> 120 -1.3863558127
#> 121 -1.2731543524
#> 122 -1.1223503200
#> 123 -0.5563430183
#> 124 -0.1413366210
#> 125  0.3312183927
#> 126  0.3850906225
#> 127  0.4395684013
#> 128  0.6032994689
#> 129  0.9497219003
#> 130  0.9227366206
#> 131  0.8966689762
#> 132  0.8706353184
#> 133  0.7794495431
#> 134  0.7012805967
#> 135  0.6165862425
#> 136  0.6403352821
#> 137  0.6786282692
#> 138  0.8340878105
#> 139 -0.1269786344
#> 140 -0.1564594022
#> 141 -0.1895175147
#> 142 -0.2308078298
#> 143 -0.3382143700
#> 144 -0.4620853155
#> 145 -0.5575912672
#> 146 -0.5445416651
#> 147 -0.4178235884
#> 148  1.0181487465
#> 149  0.8945044375
#> 150  0.8157719041
#> 151  0.6844597309
#> 152  0.3433767035
#> 153  0.0021567499
#> 154 -0.3495000971
#> 155 -0.3074940243
#> 156 -0.2633430273
#> 157 -0.0962009100
#> 158 -0.4163998955
#> 159 -0.3349026588
#> 160 -0.2797072393
#> 161 -0.1877468709
#> 162  0.0513692853
#> 163  0.3089159121
#> 164  0.5435577836
#> 165  0.5456449839
#> 166  0.5481584927
#> 167  0.7640214722
#> 168  0.1176552650
#> 169 -0.0035667671
#> 170 -0.1837217695
#> 171 -0.5696723148
#> 172 -0.9040732916
#> 173 -1.2232572857
#> 174 -1.1863970493
#> 175 -1.1276515100
#> 176 -0.2072636242
#> 177 -0.1321279401
#> 178 -0.0587381474
#> 179  0.0965838353
#> 180  0.2638776448
#> 181  0.4072901539
#> 182  0.3779461550
#> 183  0.3614578994
#> 184  0.3304820293
#> 185  1.3533699181
#> 186  1.2936701283
#> 187  1.2413987369
#> 188  1.1891954960
#> 189  1.0063478520
#> 190  0.8365169430
#> 191  0.6797709194
#> 192  0.6250163987
#> 193  0.5601970130
#> 194  0.5871094973
#> 195  0.2254709467
#> 196  0.2217392918
#> 197  0.2195756496
#> 198  0.2161382214
#> 199  0.2072001902
#> 200  0.1982585710
#> 201  0.1900058727
#> 202  0.1957485770
#> 203  0.2108115697
#> 204  0.2609467175
#> 205  0.0338941746
#> 206  0.0338015529
#> 207  0.0337450029
#> 208  0.0336550114
#> 209  0.0334029789
#> 210  0.0331869243
#> 211  0.0325909332
#> 212 -0.0105713603
#> 213 -0.0632290119
#> 214 -0.1350198498
#> 215 -0.1245228591
#> 216  0.0002658531
#> 217  0.0671606036
#> 218  0.1522791731
#> 219  0.4291923261
#> 220  0.6847702762
#> 221  0.9829075112
#> 222  0.9756432884
#> 223  0.9653721276
#> 224  0.9531145604
#> 225 -0.1257141349
#> 226 -0.1001136484
#> 227 -0.0839997722
#> 228 -0.0583992857
#> 229  0.0081673238
#> 230  0.0696298698
#> 231  0.1362232021
#> 232  0.1757273494
#> 233  0.2025084646
#> 234  0.2735462193
#> 235  0.0135962215
#> 236 -0.1544522883
#> 237 -0.4449547509
#> 238 -0.9799929189
#> 239 -1.4080234532
#> 240 -1.8432487574
#> 241 -1.7696957423
#> 242 -1.7022828051
#> 243 -1.5315438714
#> 244 -1.2543551235
#> 245 -1.1198584769
#> 246 -1.0459160283
#> 247 -0.9517069617
#> 248 -0.6221594706
#> 249 -0.3160721251
#> 250 -0.0336905813
#> 251 -0.0964562161
#> 252 -0.1626481624
#> 253 -0.3580724551
#> 254 -0.7072102380
#> 255 -0.6334261510
#> 256 -0.5775600161
#> 257 -0.5217667184
#> 258 -0.3402564077
#> 259 -0.1588189342
#> 260  0.0087066334
#> 261  0.0037028423
#> 262 -0.0164666964
#> 263  0.0031470934
#> 264  1.4930748118
#> 265  1.4686870216
#> 266  1.4442992315
#> 267  1.4207407490
#> 268  1.3501267320
#> 269  1.2677181161
#> 270  1.1815729673
#> 271  1.0646955823
#> 272  0.9899720426
#> 273  0.8531333672
#> 274  0.9057357307
#> 275  0.9603311473
#> 276  1.0148553834
#> 277  1.2194992200
#> 278  1.3832142893
#> 279  1.5605248274
#> 280  1.5100322077
#> 281  1.4496416739
#> 282  1.4910729470
#> 283  0.3454157264
#> 284  0.3014971510
#> 285  0.2717525329
#> 286  0.2165568675
#> 287  0.0876462789
#> 288 -0.0313321937
#> 289 -0.1728397109
#> 290 -0.1880413893
#> 291 -0.2196037826
#> 292 -0.1877493378
#> 293  2.3698212709
#> 294  2.3737098902
#> 295  2.3763435251
#> 296  2.3824856199
#> 297  2.3938995649
#> 298  2.4035546997
#> 299  2.4149640644
#> 300  2.4269190442
#> 301  2.4381276141
#> 302  2.4564749263
#> 303 -0.2958075032
#> 304 -0.4082687396
#> 305 -0.4757689598
#> 306 -0.5882301961
#> 307 -0.9033094846
#> 308 -1.1731929736
#> 309 -1.4657330581
#> 310 -1.4145997889
#> 311 -1.3502149378
#> 312 -0.5159301205
#> 313 -0.8329744561
#> 314 -0.8339879285
#> 315 -0.8353515613
#> 316 -0.8367151941
#> 317 -0.8405342201
#> 318 -0.8438080776
#> 319 -0.8467476800
#> 320 -0.8364593392
#> 321 -0.8255534294
#> 322 -0.7974396951
#> 323  1.3311062390
#> 324  1.3981490359
#> 325  1.4568497497
#> 326  1.5448625539
#> 327  1.6517024449
#> 328  1.7104031587
#> 329  1.9115315495
#> 330  2.0915986554
#> 331  2.0274244648
#> 332  1.8705790225
#> 333  1.3567850449
#> 334 -1.7331222690
#> 335 -1.5134468577
#> 336 -1.3375689449
#> 337 -1.2057178369
#> 338 -0.6345159064
#> 339 -0.0192871711
#> 340  0.5521440657
#> 341  0.5656038383
#> 342  0.5741230797
#> 343  0.6587121325
#> 344 -2.3777537549
#> 345 -2.1506090207
#> 346 -1.9876943519
#> 347 -1.8654020091
#> 348 -1.3353976294
#> 349 -0.8056059321
#> 350 -0.2492832099
#> 351 -0.1945525228
#> 352 -0.1109850011
#> 353 -1.9510584003
#> 354 -1.8520218519
#> 355 -1.7626673218
#> 356 -1.6285188758
#> 357 -1.3602219838
#> 358 -1.0696447847
#> 359 -0.8013478927
#> 360 -0.7633108245
#> 361 -0.7301311249
#> 362 -0.6193465020
#> 363  0.4832731398
#> 364  0.4615384608
#> 365  0.4343913649
#> 366  0.4126850231
#> 367  0.3366278151
#> 368  0.2714804525
#> 369  0.2225986777
#> 370  0.1895116133
#> 371  0.1847205143
#> 372  0.1742386179
#> 373  0.4501388384
#> 374  0.3732398223
#> 375  0.3270843585
#> 376  0.2193615196
#> 377  0.0808951284
#> 378 -0.1498019201
#> 379 -0.3439696869
#> 380 -0.2920537404
#> 381 -0.2151895711
#> 382  0.1055857536
#> 383  0.6675452430
#> 384  0.7037572471
#> 385  0.7311754181
#> 386  0.7585578418
#> 387  0.8476400870
#> 388  0.9298231053
#> 389  1.0248759172
#> 390  1.0080059475
#> 391  0.9853862369
#> 392  1.1582634595
#> 393 -0.0064460028
#> 394  0.0185454781
#> 395  0.0391077492
#> 396  0.0648508536
#> 397  0.1317185001
#> 398  0.1934590008
#> 399  0.2603266473
#> 400  0.2468366519
#> 401  0.2321563091
#> 402  0.2010806301
#> 403 -1.8226982863
#> 404 -2.0384450445
#> 405 -2.1293419210
#> 406 -2.2769456493
#> 407 -2.4132317091
#> 408 -2.4654490547
#> 409 -2.5191661288
#> 410 -2.6092122094
#> 411  0.7857963046
#> 412  0.6632127737
#> 413  0.5153097226
#> 414  0.3968946493
#> 415  0.0123158389
#> 416 -0.3724173588
#> 417 -0.7275081914
#> 418 -0.7850381064
#> 419 -0.8093954312
#> 420 -0.8348289837
#> 421 -1.0314465327
#> 422 -0.9341102828
#> 423 -0.8431761735
#> 424 -0.7522420643
#> 425 -0.4341505192
#> 426 -0.1160589741
#> 427  0.1540661251
#> 428  0.1458874619
#> 429  0.1366285977
#> 430 -0.0065995049
#> 431  1.1767672698
#> 432  1.1433378783
#> 433  1.1186837021
#> 434  1.0940295259
#> 435  1.0139597045
#> 436  0.9030641270
#> 437  0.8551791952
#> 438  0.8342270405
#> 439  0.8180183611
#> 440  0.8467679342
#> 441 -0.7152268668
#> 442 -0.7214997170
#> 443 -0.7253739263
#> 444 -0.7318286959
#> 445 -0.7486191823
#> 446 -0.7563608632
#> 447 -0.7654029308
#> 448 -0.7838101674
#> 449 -0.7902695271
#> 450 -0.8023644442
#> 451 -0.8758877590
#> 452  2.0058738165
#> 453  1.9706340497
#> 454  1.9327216039
#> 455  1.8758034412
#> 456  1.7524642574
#> 457  1.6386279320
#> 458  1.5080756228
#> 459  1.4442917949
#> 460  1.3736274149
#> 461  1.3040028507
#> 462  1.1987786822
#> 463  1.1890853980
#> 464  1.1869295864
#> 465  1.1793921138
#> 466  1.1597852434
#> 467  1.1401862410
#> 468  1.1208295500
#> 469  1.1337471082
#> 470  1.1475857445
#> 471  1.1830761912
#> 472 -0.2426800430
#> 473 -0.3820037731
#> 474 -0.4656270975
#> 475 -0.6049508276
#> 476 -0.9673670440
#> 477 -1.3296378286
#> 478 -1.6879819528
#> 479 -1.7038483062
#> 480 -1.7149157961
#> 481 -1.7246869679
#> 482 -0.0556379503
#> 483 -0.0022067792
#> 484  0.0512243919
#> 485  0.0832942492
#> 486  0.2329907662
#> 487  0.3612144221
#> 488  0.5002023953
#> 489  0.5358210964
#> 490  0.5611633408
#> 491  0.8786115566
#> 492 -0.4864417699
#> 493 -0.4418868121
#> 494 -0.3862366823
#> 495 -0.3417398144
#> 496 -0.1969797691
#> 497 -0.0522778136
#> 498  0.0949789584
#> 499  0.1013486472
#> 500  0.1139949883
#> 501  0.1272759722
#> 502 -0.0277833913
#> 503 -0.1332909552
#> 504 -0.1984540586
#> 505 -0.3070214552
#> 506 -0.5894326789
#> 507 -0.8717305754
#> 508 -1.1540284720
#> 509 -1.1527390585
#> 510 -1.1420586934
#> 511 -0.9801906919
#> 512 -0.1911340679
#> 513 -0.2295998473
#> 514 -0.2949448148
#> 515 -0.4010558410
#> 516 -0.5072094649
#> 517 -0.6133204911
#> 518 -0.5676520960
#> 519 -0.5050760294
#> 520 -0.4182437580
#> 521 -0.2428005984
#> 522 -0.1794252136
#> 523 -0.1182049111
#> 524 -0.0417393964
#> 525  0.1570869054
#> 526  0.3559930250
#> 527  0.5395741146
#> 528  0.5574278538
#> 529  0.5595279727
#> 530  0.6744285088
#> 531  0.8005878926
#> 532  0.7353000865
#> 533  0.6798452169
#> 534  0.6105808556
#> 535  0.4442885476
#> 536  0.2641144470
#> 537  0.0768621055
#> 538  0.0717301103
#> 539  0.0647308284
#> 540  0.0541333069
#> 541 -0.0408898839
#> 542 -0.1212627274
#> 543 -0.1873911895
#> 544 -0.2701812615
#> 545 -0.4687393068
#> 546 -0.6837863029
#> 547 -0.9068799096
#> 548 -0.9545461064
#> 549 -1.0101497611
#> 550 -1.1083006775
#> 551 -1.5888799038
#> 552 -1.5373884201
#> 553 -1.4909308506
#> 554 -1.4444126316
#> 555 -1.2816898392
#> 556 -1.1421958315
#> 557 -0.9506669432
#> 558 -0.9198865705
#> 559 -0.5959059753
#> 560 -0.6990209235
#> 561 -0.6730323845
#> 562 -0.6582807458
#> 563 -0.6337032331
#> 564 -0.5648708046
#> 565 -0.4763609728
#> 566 -0.4015765642
#> 567 -0.3574590995
#> 568 -0.2116345091
#> 569  0.4841238642
#> 570  0.4775003632
#> 571  0.4713125126
#> 572  0.4635837501
#> 573  0.4434792864
#> 574  0.4218339108
#> 575  0.4039518079
#> 576  0.4637312509
#> 577  0.5129994952
#> 578  0.6371900280
#> 579  0.9972611260
#> 580  1.0049630689
#> 581  1.0119120539
#> 582  1.0206118928
#> 583  1.0429647636
#> 584  1.0658075106
#> 585  1.0882837173
#> 586  1.1041000907
#> 587  1.1199228134
#> 588  1.1489523031
#> 589  0.6612163540
#> 590  0.6682576053
#> 591  0.6750594186
#> 592  0.6835550341
#> 593  0.7073480784
#> 594  0.7277446501
#> 595  0.7489861775
#> 596  0.7376611870
#> 597  0.7878323696
#> 598 -0.4535817949
#> 599 -0.3761061935
#> 600 -0.3102991211
#> 601 -0.2444920488
#> 602 -0.0142959927
#> 603  0.1830394263
#> 604  0.3967622648
#> 605  0.3991842490
#> 606  0.3956994895
#> 607  0.3959543832
#> 608 -0.3799767636
#> 609 -0.4199852624
#> 610 -0.4578215327
#> 611 -0.4881142460
#> 612 -0.5789528908
#> 613 -0.6773350926
#> 614 -0.7750497981
#> 615 -0.7763935253
#> 616 -0.7765618057
#> 617 -0.7540043642
#> 618 -0.6821219417
#> 619 -0.6172668938
#> 620 -0.5384417639
#> 621 -0.3089090814
#> 622 -0.1304306990
#> 623  0.0804752074
#> 624  0.1197568966
#> 625  0.1456581706
#> 626 -0.4481597332
#> 627 -0.4851557372
#> 628 -0.5197183866
#> 629 -0.5628879016
#> 630 -0.6665307875
#> 631 -0.7787805392
#> 632 -0.8941581409
#> 633 -0.8615808830
#> 634 -0.8224126520
#> 635 -0.7971931869
#> 636  0.7510672794
#> 637  0.6994063453
#> 638  0.6829863139
#> 639  0.6419147995
#> 640  0.5350774154
#> 641  0.4200514517
#> 642  0.3221399924
#> 643  0.2977232112
#> 644  0.2749139580
#> 645  0.2831760872
#> 646 -1.4355857021
#> 647 -1.3319608086
#> 648 -1.1976270495
#> 649 -1.0900759083
#> 650 -0.7406398669
#> 651 -0.3913440486
#> 652 -0.0420482303
#> 653 -0.0365803595
#> 654 -0.0384611874
#> 655 -0.0440167149
#> 656  0.0071996654
#> 657 -0.1048342510
#> 658 -0.2201172713
#> 659 -0.2860620470
#> 660 -0.5891914084
#> 661 -0.8889513286
#> 662 -1.1887112489
#> 663 -1.1640602834
#> 664 -1.1340363451
#> 665 -1.0601560150
#> 666  1.1378809856
#> 667  1.0832519073
#> 668  1.0265529857
#> 669  0.9557347759
#> 670  0.7715187229
#> 671  0.5873765928
#> 672  0.4032344628
#> 673  0.3886252229
#> 674  0.3815025824
#> 675  0.3617068442
#> 676 -2.9755925121
#> 677 -2.7815385147
#> 678 -2.6084838417
#> 679 -2.3661060980
#> 680 -1.6872966131
#> 681 -1.0568108746
#> 682 -0.4442382553
#> 683 -0.4039061355
#> 684 -0.3362183170
#> 685 -0.2752577417
#> 686 -0.3411737042
#> 687 -0.3975930740
#> 688 -0.4520939011
#> 689 -0.5181770421
#> 690 -0.6971699799
#> 691 -0.8742443749
#> 692 -1.0512477127
#> 693 -0.9765700958
#> 694 -0.9195933988
#> 695 -0.8142970140
#> 696  0.5943051837
#> 697  0.6687736393
#> 698  0.7099461027
#> 699  0.7786146113
#> 700  0.9569808832
#> 701  1.1353471551
#> 702  1.3000370087
#> 703  1.3241129559
#> 704  1.3346521542
#> 705  1.6398542188
#> 706 -0.2895290417
#> 707 -0.2152281029
#> 708 -0.1575054669
#> 709 -0.1018174407
#> 710  0.0880041261
#> 711  0.4741456348
#> 712  0.4842855178
#> 713 -0.2877858504
#> 714 -0.2292221311
#> 715 -0.1779454499
#> 716 -0.1138997386
#> 717  0.0398634514
#> 718  0.2063956716
#> 719  0.3814779626
#> 720  0.4038095597
#> 721  0.4388992954
#> 722  0.5504863767
#> 723  1.5774912714
#> 724  1.5543866868
#> 725  1.5341569876
#> 726  1.5088896451
#> 727  1.4381252608
#> 728  1.3774625385
#> 729  1.3140300151
#> 730  1.3235479593
#> 731  1.3352870147
#> 732  1.3861183252
#> 733 -1.7938336443
#> 734 -1.8006729107
#> 735 -1.8072796076
#> 736 -1.8138776908
#> 737 -1.8353429954
#> 738 -1.8584535140
#> 739 -1.8782649910
#> 740 -1.9207964898
#> 741 -1.9758839779
#> 742 -2.0543667341

# case 1: x as knots
z <- predict(fit, test, x = "knots")

# case 2: x and y, one new group
predict(fit, test, x = "knots", y = c(1, 1, 0.5, 0))

#>   .source age hgt.z id     .pred
#> 1   added   0   1.0  0 0.8758366
#> 2   added   1   1.0  0 0.7024850
#> 3   added   2   0.5  0 0.5955456
#> 4   added   3   0.0  0 0.0729350

# case 2: x and y, one new group, we need not specify new_data
predict(fit, x = "knots", y = c(1, 1, 0.5, 0))

#>   .source age hgt.z id     .pred
#> 1   added   0   1.0  0 0.8758366
#> 2   added   1   1.0  0 0.7024850
#> 3   added   2   0.5  0 0.5955456
#> 4   added   3   0.0  0 0.0729350

# case 3: only group
predict(fit, test, group = c(11045, 11120, 999))

#> # A tibble: 20 x 9
#>    .source    id    age sex       ga    bw   hgt  hgt.z  .pred
#>    <chr>   <dbl>  <dbl> <chr>  <dbl> <dbl> <dbl>  <dbl>  <dbl>
#>  1 data    11045 0      female    40  3350  NA   NA     -0.261
#>  2 data    11045 0.0684 female    40  3350  53   -0.167 -0.285
#>  3 data    11045 0.145  female    40  3350  56   -0.028 -0.313
#>  4 data    11045 0.241  female    40  3350  58.5 -0.384 -0.347
#>  5 data    11045 0.548  female    40  3350  65   -1.04  -0.457
#>  6 data    11045 0.835  female    40  3350  70.5 -0.811 -0.561
#>  7 data    11045 1.06   female    40  3350  74.5 -0.627 -0.605
#>  8 data    11045 1.26   female    40  3350  77.5 -0.517 -0.559
#>  9 data    11045 1.57   female    40  3350  81   -0.626 -0.485
#> 10 data    11045 2.03   female    40  3350  88    0.073 -0.358
#> 11 data    11120 0      male      39  2780  NA   NA     -1.79 
#> 12 data    11120 0.0794 male      39  2780  52   -1.18  -1.80 
#> 13 data    11120 0.156  male      39  2780  53   -2.08  -1.81 
#> 14 data    11120 0.233  male      39  2780  55   -2.40  -1.81 
#> 15 data    11120 0.482  male      39  2780  63   -1.84  -1.84 
#> 16 data    11120 0.750  male      39  2780  68   -1.77  -1.86 
#> 17 data    11120 0.980  male      39  2780  72   -1.55  -1.88 
#> 18 data    11120 1.23   male      39  2780  74.5 -1.82  -1.92 
#> 19 data    11120 1.54   male      39  2780  76   -2.50  -1.98 
#> 20 data    11120 1.98   male      39  2780  81   -2.35  -2.05 

# case 4: predict at x in selected groups
predict(fit, test, x = c(0.5, 1, 1.25), group = c(11045, 11120, 999))

#> # A tibble: 6 x 9
#>   .source    id   age sex      ga    bw   hgt hgt.z  .pred
#>   <chr>   <dbl> <dbl> <chr> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1 added   11045  0.5  NA       NA    NA    NA    NA -0.440
#> 2 added   11045  1    NA       NA    NA    NA    NA -0.620
#> 3 added   11045  1.25 NA       NA    NA    NA    NA -0.561
#> 4 added   11120  0.5  NA       NA    NA    NA    NA -1.84 
#> 5 added   11120  1    NA       NA    NA    NA    NA -1.88 
#> 6 added   11120  1.25 NA       NA    NA    NA    NA -1.92 

# case 5: vectorized
predict(fit, test, x = c(0.5, 1, 1.25), y = c(0, 0.5, 1), group = c(11045, 11120, 999))

#> # A tibble: 3 x 9
#>   .source    id   age sex      ga    bw   hgt hgt.z  .pred
#>   <chr>   <dbl> <dbl> <chr> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1 added   11045  0.5  NA       NA    NA    NA   0   -0.389
#> 2 added   11120  1    NA       NA    NA    NA   0.5 -1.40 
#> 3 added     999  1.25 NA       NA    NA    NA   1    0.798

# case 6: vectorized, without new_data, results are different for 11045 and 11120
predict(fit, x = c(0.5, 1, 1.25), y = c(0, 0.5, 1), group = c(11045, 11120, 999))

#>   .source  age hgt.z    id        .pred
#> 1   added 0.50   0.0 11045 0.0009576034
#> 2   added 1.00   0.5 11120 0.4030588876
#> 3   added 1.25   1.0   999 0.7977639528