Predict from a brokenstick model

The predictions from a broken stick model coincide with the group-conditional means of the random effects. This function takes an object of class brokenstick and returns predictions in one of several formats. The user can calculate predictions for new persons, i.e., for persons who are not part of the fitted model, through the x and y arguments.

# S3 method for brokenstick
predict(
  object,
  newdata = NULL,
  ...,
  x = NULL,
  y = NULL,
  group = NULL,
  strip_data = TRUE,
  shape = c("long", "wide", "vector"),
  what = NULL
)

Arguments

object
A brokenstick object.

newdata
Optional. A data frame in which to look for variables with which to predict. The training data are used if omitted and if object$light is FALSE.

...
Not used, but required for extensibility.

x
Optional. A numeric vector with values of the predictor. It could also be the special keyword x = "knots" replaces x by the positions of the knots.

y
Optional. A numeric vector with measurements.

group
A vector with group identifications

strip_data
A logical indicating whether the row with the observed data from newdata should be stripped from the return value. The default is TRUE. Set to FALSE to infer which data points are extracted from newdata. Works best for shape = "long".

shape
A string: "long" (default), "wide" or "vector" specifying the shape of the return value. Note that use of "wide" with many unique values in x creates an unwieldy, large and sparse matrix.

what
Which knots to predict when x = "knots"? See get_knots(). The default, NULL, calculates all knots.

Value

If shape == "long" a long tibble of predictions. If x, y and group are not specified, the number of rows in the tibble is guaranteed to be the same as the number of rows in newdata.

If shape == "wide" a wide tibble of prediction, one record per group. Note that this format could be inefficient, depending on the data.

If shape == "vector" a vector of predicted values, of all x-values and groups.

Details

By default, predict() calculates predictions for every row in newdata. If the user specifies no newdata argument, then the function searches object for the training data (which are only available if object$light is FALSE). It is possible to tailor the behaviour of predict() through the x, y and group arguments. What exactly happens depends on which of these arguments is specified:

If the user specifies x, but no y and group, the function returns - for every group in newdata - predictions at the specified x values. This method will use the data from newdata.

If the user specifies x and y but no group, the function forms a hypothetical new group with the x and y values. This method uses no information from newdata, and also works for a light brokenstick object.

If the user specifies group, but no x or y, the function searches for the relevant data in newdata and limits its predictions to those groups. This is useful if the user needs a prediction for only one or a few groups. This does not work for a light brokenstick object.

If the user specifies x and group, but no y, the function will create new values for x in each group, search for the relevant data in newdata and provide predictions at values of x in those groups.

If the user specifies x, y and group, the function assumes that these vectors contain additional data on top on what is already available in newdata. The lengths of x, y and group must match. For a light brokenstick object, case effectively becomes case 6. See below.

As case 5, but now without newdata available. All data are specified through x, y and group and form a data frame. Matching to newdata is attempted, but as long as group id's are different from the training sample effectively new cases will be made.

Examples