Predict from a brokenstick model

The predictions from a broken stick model coincide with the group-conditional means of the random effects. This function takes an object of class brokenstick and returns predictions in one of several formats. The user can calculate predictions for new persons, i.e., for persons who are not part of the fitted model, through the x and y arguments.

# S3 method for brokenstick
predict(
  object,
  new_data = NULL,
  type = "numeric",
  ...,
  x = NULL,
  y = NULL,
  group = NULL,
  strip_data = TRUE,
  shape = c("long", "wide", "vector")
)

Arguments

object
A brokenstick object.

new_data
A data frame or matrix of new predictors.

type
A single character. The type of predictions to generate. Valid options are:

"numeric" for numeric predictions.

...
Not used, but required for extensibility.

x
Optional. A numeric vector with values of the predictor. It could also be the special keyword x = "knots" replaces x by the positions of the knots.

y
Optional. A numeric vector with measurements.

group
A vector with group identifications

strip_data
A logical indicating whether the row with the observed data from new_data should be stripped from the return. The default is TRUE. Set to FALSE to infer which data points are extracted from new_data.

shape
A string: "long" (default), "wide" or "vector" specifying the shape of the return value. Note that use of "wide" with many unique values in x creates an unwieldy, large and sparse matrix.

Value

A tibble of predictions. If x, y and group are not specified, the number of rows in the tibble is guaranteed to be the same as the number of rows in new_data.

Details

By default, predict() calculates predictions for every row in new_data. It is possible to tailor the behavior through the x, y and group arguments. What exactly happens depends on which of these arguments is specified:

If the user specifies x, but no y and group, the function returns - for every group in new_data - predictions at x values. This method will use the data from new_data.

If the user specifies x and y but no group, the function forms a hypothetical new group with the x and y values. This method uses no information from new_data.

If the user specifies group, but no x or y, the function searches for the relevant data in new_data and limits its predictions to the specified groups. This is useful if prediction for only one or a few groups is needed.

If the user specifies x and group, but no y, the function will create new values for x in each group, search for the relevant data in new_data and limit prediction to locations x in those groups.

If the user specifies x, y and group, the functions assumes that these vectors form a data frame. The lengths of x, y and group must be the same. This procedure uses only information from newdata for groups with group values that match those on newdata.

Examples

train <- smocc_200[1:1198, ]
test <- smocc_200[1199:1940, ]

# Fit
fit <- brokenstick(hgt.z ~ age | id, data = train, knots = 0:3)

# Predict, with preprocessing
predict(fit, test)

#> # A tibble: 742 x 1
#>    .pred
#>    <dbl>
#>  1 0.319
#>  2 0.325
#>  3 0.332
#>  4 0.342
#>  5 0.362
#>  6 0.387
#>  7 0.406
#>  8 0.386
#>  9 0.356
#> 10 0.322
#> # … with 732 more rows

# case 1: x as knots
z <- predict(fit, test, x = "knots")

# case 2: x and y, one new group
predict(fit, test, x = "knots", y = c(1, 1, 0.5, 0))

#>   .source age hgt.z id     .pred
#> 1   added   0   1.0  0 0.8758366
#> 2   added   1   1.0  0 0.7024850
#> 3   added   2   0.5  0 0.5955456
#> 4   added   3   0.0  0 0.0729350

# case 2: x and y, one new group, we need not specify new_data
predict(fit, x = "knots", y = c(1, 1, 0.5, 0))

#>   .source age hgt.z id     .pred
#> 1   added   0   1.0  0 0.8758366
#> 2   added   1   1.0  0 0.7024850
#> 3   added   2   0.5  0 0.5955456
#> 4   added   3   0.0  0 0.0729350

# case 3: only group
predict(fit, test, group = c(11045, 11120, 999))

#> # A tibble: 0 x 9
#> # … with 9 variables: .source <chr>, id <dbl>, age <dbl>, sex <chr>, ga <dbl>,
#> #   bw <dbl>, hgt <dbl>, hgt.z <dbl>, .pred <dbl>

# case 4: predict at x in selected groups
predict(fit, test, x = c(0.5, 1, 1.25), group = c(11045, 11120, 999))

#> # A tibble: 6 x 9
#>   .source    id   age sex      ga    bw   hgt hgt.z  .pred
#>   <chr>   <dbl> <dbl> <chr> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1 added   11045  0.5  <NA>     NA    NA    NA    NA -0.440
#> 2 added   11045  1    <NA>     NA    NA    NA    NA -0.620
#> 3 added   11045  1.25 <NA>     NA    NA    NA    NA -0.561
#> 4 added   11120  0.5  <NA>     NA    NA    NA    NA -1.84 
#> 5 added   11120  1    <NA>     NA    NA    NA    NA -1.88 
#> 6 added   11120  1.25 <NA>     NA    NA    NA    NA -1.92 

# case 5: vectorized
predict(fit, test, x = c(0.5, 1, 1.25), y = c(0, 0.5, 1), group = c(11045, 11120, 999))

#> # A tibble: 3 x 9
#>   .source    id   age sex      ga    bw   hgt hgt.z  .pred
#>   <chr>   <dbl> <dbl> <chr> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1 added   11045  0.5  <NA>     NA    NA    NA   0   -0.389
#> 2 added   11120  1    <NA>     NA    NA    NA   0.5 -1.40 
#> 3 added     999  1.25 <NA>     NA    NA    NA   1    0.798

# case 5: vectorized, without new_data, results are different for 11045 and 11120
predict(fit, x = c(0.5, 1, 1.25), y = c(0, 0.5, 1), group = c(11045, 11120, 999))

#>   .source  age hgt.z    id        .pred
#> 1   added 0.50   0.0 11045 0.0009576034
#> 2   added 1.00   0.5 11120 0.4030588876
#> 3   added 1.25   1.0   999 0.7977639528